# Use JSON source generator