--- a/src/TimeZoneNames.DataBuilder/TimeZoneNames.DataBuilder.csproj
+++ b/src/TimeZoneNames.DataBuilder/TimeZoneNames.DataBuilder.csproj
@@ -13,7 +13,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Compile Include="..\TimeZoneNames\CaseInsensitiveDictionaryConverter.cs" Link="CaseInsensitiveDictionaryConverter.cs" />
     <Compile Include="..\TimeZoneNames\TimeZoneData.cs" Link="TimeZoneData.cs" />
     <Compile Include="..\TimeZoneNames\TimeZoneValues.cs" Link="TimeZoneValues.cs" />
   </ItemGroup>