--- a/src/TimeZoneNames/IsExternalInit.cs
+++ b/src/TimeZoneNames/IsExternalInit.cs
@@ -1,8 +1,0 @@
-﻿#if !NET6_0_OR_GREATER
-
-// ReSharper disable once CheckNamespace
-namespace System.Runtime.CompilerServices;
-
-internal static class IsExternalInit { }
-
-#endif