--- a/src/TimeZoneNames/TimeZoneNames.csproj
+++ b/src/TimeZoneNames/TimeZoneNames.csproj
@@ -16,6 +16,7 @@
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <AssemblyOriginatorKeyFile>mj1856_oss.snk</AssemblyOriginatorKeyFile>
     <SignAssembly>true</SignAssembly>
+    <Trimmable>true</Trimmable>
   </PropertyGroup>
 
   <ItemGroup>