With the switch to `System.Text.Json` in #85 we can further improve TimeZoneNames for NativeAOT applications by switching to a source generated `TimeZoneData` JSON deserializer.

Unfortunately `System.Text.Json` does not (yet) support init only properties, see https://github.com/dotnet/runtime/issues/58770.
For this draft I went with public setters, but alternatively a parameterized constructor should also work.

When I ran `DataBuilder` the order of the output changed, which I _guess_ is either due [non-stable dictionary sorting](https://github.com/dotnet/runtime/issues/76008) or that the CurrentCulture on my machine is a mix of en-US/en-GB and da-DK (reported as en-DK).

I made a quick benchmark on calling `TimeZoneData.Load()` on .NET6 using Benchmarkdotnet which _only_ focuses on the performance benefits of switching to a source generated JSON deserializer.
| branch |     Mean |   Error |  StdDev | Ratio |
|------- |---------:|--------:|--------:|--------:|
|   main | 174.5 ms | 2.98 ms | 2.49 ms | 1.00 |
|   PR | 114.8 ms | 2.25 ms | 2.10 ms | 0.66 |

Some stats on a [`TimeZoneNames.NativeAot`](https://github.com/jnyrup/TimeZoneNames/commit/9d55806aa0a7a5ba3ec5748a1fa6906d3c364158) application.

main:
* File size of TimeZoneNames.NativeAot.exe is 6091 KB
* Crashes at runtime with 
```
Unhandled Exception: System.TypeInitializationException: TypeInitialization_Type_NoTypeAvailable
	---> System.NotSupportedException: DeserializeNoConstructor, JsonConstructorAttribute, TimeZoneNames.TimeZoneData Path: $ | LineNumber: 0 | BytePositionInLine: 1.
	---> System.NotSupportedException: DeserializeNoConstructor, JsonConstructorAttribute, TimeZoneNames.TimeZoneData
```

PR:
* File size of TimeZoneNames.NativeAot.exe is 4420 KB
* Works at runtime.